Fix alignment with -l option

+++ libft/inc/libft_printf.h
@@ -50,7 +50,6 @@ typedef struct {
 	PrintfSpecifier	specifier;
 }	PrintfFormat;
 
-// PRINTF
 int	ft_printf(const char *s, ...);
 void ft_printf_parse_format(const char *s, int *i, va_list *params, PrintfFormat *fmt);
 
@@ -59,7 +58,7 @@ int	ft_putchar_formatted(char c, PrintfFormat *fmt);
 int	ft_puthex_formatted(unsigned long long n, PrintfFormat *fmt, char format);
 int	ft_puthex(unsigned long long n, char format);
 int	ft_putnbr_formatted(long long n, PrintfFormat *fmt);
-int	ft_putunbr_formatted(unsigned int n, PrintfFormat *fmt);
+int	ft_putunbr_formatted(unsigned long long n, PrintfFormat *fmt);
 int	ft_putptr_formatted(unsigned long long n, PrintfFormat *fmt);
 int ft_putoct_formatted(unsigned long long n, PrintfFormat *fmt);
